Se crea el policy para branch ademas de crear e implementar nuevas excepciones en los casos de uso

+++ modules/Iam/Branch/Application/Branch/UseCases/DeleteBranchByIdUseCase.php
@@ -2,6 +2,8 @@
 
 namespace Modules\Iam\Branch\Application\Branch\UseCases;
 
+use Modules\Iam\Branch\Application\Branch\Exceptions\CannotDeleteMainBranchException;
+use Modules\Iam\Branch\Application\Exceptions\BranchNotFoundException;
 use Modules\Iam\Branch\Domain\Repositories\BranchRepository;
 
 class DeleteBranchByIdUseCase
@@ -12,6 +14,16 @@ class DeleteBranchByIdUseCase
 
     public function execute(int $id): void
     {
+        $branch = $this->repository->findById($id);
+
+        if ($branch === null) {
+            throw new BranchNotFoundException();
+        }
+
+        if ($branch->isMain()) {
+            throw new CannotDeleteMainBranchException();
+        }
+
         $this->repository->deleteById($id);
     }
 }
